Move a few tests
