+++ test/TemplateTest.js
@@ -593,70 +593,6 @@ test("Clone the template", async (t) => {
   t.is(cloned.extensionMap, tmpl.extensionMap);
 });
 
-test("Permalink with variables!", async (t) => {
-  let tmpl = getNewTemplate(
-    "./test/stubs/permalinkdata.njk",
-    "./test/stubs/",
-    "./dist"
-  );
-
-  t.is(await tmpl.getOutputPath(), "./dist/subdir/slug-candidate/index.html");
-});
-
-test("Permalink with variables and JS front matter!", async (t) => {
-  let tmpl = getNewTemplate(
-    "./test/stubs/permalinkdata-jsfn.njk",
-    "./test/stubs/",
-    "./dist"
-  );
-
-  t.is(await tmpl.getOutputPath(), "./dist/subdir/slug/index.html");
-});
-
-// This is broken right now, permalink must use the same template language as the template
-test.skip("Use a JavaScript function for permalink in any template language", async (t) => {
-  let tmpl = getNewTemplate(
-    "./test/stubs/permalinkdata-jspermalinkfn.njk",
-    "./test/stubs/",
-    "./dist"
-  );
-
-  t.is(await tmpl.getOutputPath(), "./dist/subdir/slug/index.html");
-});
-
-test("Permalink with dates!", async (t) => {
-  let tmpl = getNewTemplate(
-    "./test/stubs/permalinkdate.liquid",
-    "./test/stubs/",
-    "./dist"
-  );
-
-  t.is(await tmpl.getOutputPath(), "./dist/2016/01/01/index.html");
-});
-
-test("Permalink with dates on file name regex!", async (t) => {
-  let tmpl = getNewTemplate(
-    "./test/stubs/2016-02-01-permalinkdate.liquid",
-    "./test/stubs/",
-    "./dist"
-  );
-
-  t.is(await tmpl.getOutputPath(), "./dist/2016/02/01/index.html");
-});
-
-test("Reuse permalink in directory specific data file", async (t) => {
-  let eleventyConfig = new TemplateConfig();
-  let dataObj = new TemplateData("./test/stubs/", eleventyConfig);
-  let tmpl = getNewTemplate(
-    "./test/stubs/reuse-permalink/test1.liquid",
-    "./test/stubs/",
-    "./dist",
-    dataObj
-  );
-
-  t.is(await tmpl.getOutputPath(), "./dist/2016/01/01/index.html");
-});
-
 test("mapDataAsRenderedTemplates", async (t) => {
   let tmpl = getNewTemplate(
     "./test/stubs/default.ejs",
